[REFACTOR] change to long from int

+++ Analyzer Service/Models/Schema/HistoricalSimilarityPoint.cs
@@ -13,10 +13,10 @@ namespace Analyzer_Service.Models.Schema
         public int ComparedFlightIndex { get; set; }
 
         [BsonElement("StartIndex")]
-        public int StartIndex { get; set; }
+        public long StartIndex { get; set; }
 
         [BsonElement("EndIndex")]
-        public int EndIndex { get; set; }
+        public long EndIndex { get; set; }
 
         [BsonElement("Label")]
         public string Label { get; set; } = string.Empty;
